move ScanEnum() to new file
